// src/state/actionType/action-types.ts
export enum STORE_ActionType {
    ADD_STORE = "ADD_STORE",
    DELETE_STORE = "DELETE_STORE",
    UPDATE_STORE = "UPDATE_STORE",
    VIEW_STORE = "VIEW_STORE",
    SEARCH_STORE = "SEARCH_STORE"
} 
export enum PRODUCT_ActionType {
    ADD_PRODUCT = "ADD_PRODUCT",
    DELETE_PRODUCT = "DELETE_PRODUCT",
    UPDATE_PRODUCT = "UPDATE_PRODUCT",
    VIEW_PRODUCT = "VIEW_PRODUCT",
    SEARCH_PRODUCT = "SEARCH_PRODUCT"
} 